+++ services/workers/impl/src/main/kotlin/ss/workers/impl/di/WorkersModule.kt
@@ -13,22 +13,35 @@
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
  * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
 
-package ss.workers.api
+package ss.workers.impl.di
 
-/** API to schedule background work. **/
-interface WorkScheduler {
+import android.content.Context
+import androidx.work.WorkManager
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import ss.workers.api.WorkScheduler
+import ss.workers.impl.WorkSchedulerImpl
+import javax.inject.Singleton
 
-    /**
-     * Prefetch cover images for [language] quarterlies.
-     *
-     * @param language The current selected language.
-     */
-    fun preFetchImages(language: String)
+@Module
+@InstallIn(SingletonComponent::class)
+object WorkersModule {
+
+    @Provides
+    @Singleton
+    fun provideWorkSchedulerApi(
+        @ApplicationContext context: Context
+    ): WorkScheduler = WorkSchedulerImpl(
+        workManager = WorkManager.getInstance(context.applicationContext)
+    )
 }
